Add files via upload
Final	Project Assignment
CS590-01	Summer	2016

The	assignment will be	submitted through	Canvas.
As	before,	you	need	include a	Makefile,	then	tar and	gzip	your	files	and	upload	the	gz	file.
Goal: In this assignment,	you	will	practice using the	pthread library	to write a	C program that	creates a	number	of	producer	and	consumer	threads. The producer threads will	generate	random	numbers	to be stored in a circular queue. The consumer threads	 will remove elements from the queue.
Assignment	Details
1. The program must ask	the	user	for	the	number	of	producer	and	number	of	
consumer	threads to create.
a. The number	of	producers and	consumers do	not	have	to	be	balanced.	In	
other words, there	may	be	50	producers and	1 consumer,	or 50 consumers	
and	10	producers, and so on. So, be	sure	to test in	a variety	of ways.
2. Producers and consumers	will operate on	a	circular	queue.
a. The	queue	must	have	a	capacity	of 100	integers.	
b. You	can	search	for	information	on circular	queues	if	yo